integration-20190725 (#534)
* remove unused code stub

* ASD-144: remove debugging line

* ASD-144: preserve backward-compatibility of AmazonAddressFactory constructor

* ASD-144: move disable-login-layout logic to a separate class

* ASD-144: don't rely on order of getStores() return value

* Increase version to 3.2.11

* revert ASD-50: move methods back to Core\Helper\Data

* ASD-147: update PHP requirement in composer.json

* ASD-132: add missing javascript dependencies

* ASD-155: fix renumbering of simplepath field indices
