+++ composer.json
@@ -11,7 +11,7 @@
     "phpunit/phpunit": "4.1.0"
   },
   "require": {
-    "php": "~7.1.3||~7.2.0",
+    "php": "~7.1.3||~7.2.0||~7.3.0",
     "magento/framework": "^102",
     "magento/module-sales": "^102",
     "magento/module-checkout": "^100.3",
