+++ src/Payment/view/frontend/web/js/amazon-core.js
@@ -18,10 +18,11 @@ define([
     'ko',
     'mage/url',
     'amazonPaymentConfig',
+    'Magento_Ui/js/model/messageList',
     'amazonWidgetsLoader',
     'jquery/jquery-storageapi',
     'mage/cookies'
-], function ($, ko, url, amazonPaymentConfig) {
+], function ($, ko, url, amazonPaymentConfig, messageList) {
     'use strict';
 
     var amazonDefined = ko.observable(false),
@@ -54,13 +55,6 @@ define([
     window.onAmazonPaymentsReady = function() {
         $(window).trigger('OffAmazonPayments');
     };
-
-    /**
-     * Set Client ID
-     * @param {String} cid
-     */
-    function setClientId(cid) {
-    }
 
     /**
      * Log user out of amazon
@@ -107,11 +101,12 @@ define([
     }
 
     function handleWidgetError(error) {
-        console.log('OffAmazonPayments.Widgets.AddressBook', error.getErrorCode(), error.getErrorMessage());
+        console.log('OffAmazonPayments.Widgets', error.getErrorCode(), error.getErrorMessage());
         switch (error.getErrorCode()) {
             case 'BuyerSessionExpired':
                 messageList.addErrorMessage({message: $.mage.__('Your Amazon session has expired.  Please sign in again by clicking the Amazon Pay Button.')});
-                amazonStorage.amazonlogOut();
+                var storage = require('Amazon_Payment/js/model/storage'); //TODO: clean up this circular dependency
+                storage.amazonlogOut();
                 break;
             case 'ITP':
                 // ITP errors are how handled within the widget code
